tests/wpt/web-platform-tests/workers/semantics/encodings/001.html

<!--
postMessage('å');
/*
-->
<!doctype html>
<title>encoding, dedicated worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, '\u00e5');
    this.done();
  });
});
</script>
<!--
*/
//-->
